backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Pillow==10.0.1
tensorflow==2.18.0
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2 